перенес код для рендера в отдельный модуль, исправил пути
