+++ src/components/abstract-component.js
@@ -1,4 +1,4 @@
-import {utils} from "../utils.js";
+import {createElement} from "../utils/render.js";
 
 export default class AbstractComponent {
   constructor() {
@@ -15,7 +15,7 @@ export default class AbstractComponent {
 
   getElement() {
     if (!this._element) {
-      this._element = utils.createElement(this.getTemplate());
+      this._element = createElement(this.getTemplate());
     }
 
     return this._element;
